requirements.txt
pyttsx3
speechrecognition
pyaudio
google-generativeai
